Add a more robust ticket preset system

+++ src/interactions/buttons/Support Ticket/createSupportTicketWithPreset.ts
@@ -2,11 +2,9 @@ import { DatabaseManager } from "@alice-database/DatabaseManager";
 import { CreateSupportTicketWithPresetLocalization } from "@alice-localization/interactions/buttons/Support Ticket/createSupportTicketWithPreset/CreateSupportTicketWithPresetLocalization";
 import { ButtonCommand } from "@alice-structures/core/ButtonCommand";
 import { MessageCreator } from "@alice-utils/creators/MessageCreator";
-import { ModalCreator } from "@alice-utils/creators/ModalCreator";
 import { SelectMenuCreator } from "@alice-utils/creators/SelectMenuCreator";
 import { CommandHelper } from "@alice-utils/helpers/CommandHelper";
 import { InteractionHelper } from "@alice-utils/helpers/InteractionHelper";
-import { TextInputBuilder, TextInputStyle } from "discord.js";
 
 export const run: ButtonCommand["run"] = async (_, interaction) => {
     const dbManager = DatabaseManager.aliceDb.collections.supportTicketPreset;
@@ -53,12 +51,9 @@ export const run: ButtonCommand["run"] = async (_, interaction) => {
         return;
     }
 
-    const preset = await dbManager.getOne(
-        {
-            id: parseInt(selectMenuInteraction.values[0]),
-        },
-        { projection: { _id: 0, title: 1, description: 1 } },
-    );
+    const preset = await dbManager.getOne({
+        id: parseInt(selectMenuInteraction.values[0]),
+    });
     if (!preset) {
         return InteractionHelper.reply(interaction, {
             content: MessageCreator.createReject(
@@ -67,31 +62,11 @@ export const run: ButtonCommand["run"] = async (_, interaction) => {
         });
     }
 
-    await ModalCreator.createModal(
-        selectMenuInteraction,
-        "ticket-create",
-        localization.getTranslation("modalTitle"),
-        new TextInputBuilder()
-            .setCustomId("title")
-            .setRequired(true)
-            .setStyle(TextInputStyle.Short)
-            .setMaxLength(100)
-            .setPlaceholder(
-                localization.getTranslation("modalTitlePlaceholder"),
-            )
-            .setLabel(localization.getTranslation("modalTitleLabel"))
-            .setValue(preset.title),
-        new TextInputBuilder()
-            .setCustomId("description")
-            .setRequired(true)
-            .setStyle(TextInputStyle.Paragraph)
-            .setMaxLength(1500)
-            .setPlaceholder(
-                localization.getTranslation("modalDescriptionPlaceholder"),
-            )
-            .setLabel(localization.getTranslation("modalDescriptionLabel"))
-            .setValue(preset.description),
-    );
+    if (!(await preset.validate(selectMenuInteraction))) {
+        return;
+    }
+
+    await selectMenuInteraction.showModal(preset.buildModal(language));
 
     // Delete the left-over select menu, which is a reply to the original interaction.
     await interaction.deleteReply();
